@@ -21,8 +21,7 @@
     "@chakra-ui/react": "^2.0.2",
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
-    "@fontsource/inter": "^4.5.10",
-    "@fontsource/raleway": "^4.5.8",
+    "@fontsource/plus-jakarta-sans": "^4.5.8",
     "framer-motion": "^6.3.3",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
@@ -38,7 +37,7 @@
     "@types/react-helmet": "^6.1.5",
     "@vitejs/plugin-react": "^1.3.2",
     "commitlint": "^17.0.0",
-    "eslint": "^8.15.0",
+    "eslint": "^8.16.0",
     "eslint-config-sznm": "^1.0.0",
     "husky": "^8.0.1",
     "lint-staged": "^12.4.1",
