@@ -2,8 +2,8 @@ import { extendTheme } from "@chakra-ui/react";
 
 export const theme = extendTheme({
   fonts: {
-    heading: "Raleway, sans-serif",
-    body: "Inter, sans-serif",
+    heading: "Plus Jakarta Sans, sans-serif",
+    body: "Plus Jakarta Sans, sans-serif",
   },
   components: {
     // Button: {
